[Zurich] Set report's public response before sending email to user
The 'public response' to a problem was being omitted from the email sent to the
user when the problem is closed, because the email was sent before the report's
'extra' field was updated with the response body text.

This commit sets the public response on the problem before sending the email.